+++ rule_shellcheck_test.go
@@ -62,6 +62,10 @@ func TestRuleShellcheckSanitizeExpressionsInScript(t *testing.T) {
 			"a}}b${{c}}d",
 			"a}}b______d",
 		},
+		{
+			"before ${{\nfoo\n}} after",
+			"before ___\n___\n__ after",
+		},
 	}
 
 	for i, tc := range testCases {